backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
websockets==12.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Redis & RabbitMQ
redis==5.0.1
aio-pika==9.3.1

# Machine Learning
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2
scikit-learn==1.3.2

# Vector Database
chromadb==0.4.22

# OpenTelemetry
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0

# Development
pytest==7.4.3
black==23.12.1
ruff==0.1.9
